+++ Craftsman/Builders/Features/ConsumerBuilder.cs
@@ -29,9 +29,9 @@
         {
             var context = Utilities.GetDbContext(srcDirectory, projectBaseName);
             var contextClassPath = ClassPathHelper.DbContextClassPath(srcDirectory, "", projectBaseName);
-            var dbReadOnly = consumer.UsesDb ? @$"{Environment.NewLine}        private readonly {context} _db;" : "";
+            var dbReadOnly = consumer.UsesDb ? @$"{Environment.NewLine}    private readonly {context} _db;" : "";
             var dbProp = consumer.UsesDb ? @$"{context} db, " : "";
-            var assignDb = consumer.UsesDb ? @$"{Environment.NewLine}            _db = db;" : "";
+            var assignDb = consumer.UsesDb ? @$"{Environment.NewLine}        _db = db;" : "";
             var contextUsing = consumer.UsesDb ? $@"
     using {contextClassPath.ClassNamespace};" : "";
 
@@ -49,14 +49,6 @@ public class {consumer.ConsumerName} : IConsumer<{consumer.MessageName}>
     public {consumer.ConsumerName}({dbProp}IMapper mapper)
     {{
         _mapper = mapper;{assignDb}
-    }}
-
-    public class {consumer.ConsumerName}Profile : Profile
-    {{
-        public {consumer.ConsumerName}Profile()
-        {{
-            //createmap<to this, from this>
-        }}
     }}
 
     public Task Consume(ConsumeContext<{consumer.MessageName}> context)